List Class Implemented
I've integrated a List class that will store all of the receipts for
each person within itself.  It will also keep track of when a Receipt
List has been completed so that duplicate work is avoided.

Steps 4-6 still need to be implemented.  I will work on those when I get
a chance to sit down with this again.